Disable building of the examples
`libzt` adds build target for its examples by default.

However, the `examples/` directory is marked as `export-ignore` in libzt's
`.gitattributes`, so it does not get included into the GitHub-provided archive.

While we use a recursive git fetch (and are not affected), distributions
may override this with `-DFETCHCONTENT_SOURCE_DIR_LIBZT=...`,
where `export-ignore` files may be omitted from the source directory.